% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_report.R
\name{create_vrule_report}
\alias{create_vrule_report}
\title{create_vrule_report}
\usage{
create_vrule_report(valid, category, rule, type, message)
}
\arguments{
\item{valid}{valid \code{TRUE} or \code{FALSE}}

\item{category}{category}

\item{rule}{rule}

\item{type}{type}

\item{message}{message}
}
\description{
create_vrule_report
}
