% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_options.R
\name{getVruleOption}
\alias{getVruleOption}
\title{getVruleOption}
\usage{
getVruleOption(opt)
}
\arguments{
\item{opt}{an option name}
}
\value{
the value for that option
}
\description{
getVruleOption
}
