% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule.R
\docType{package}
\name{vrule}
\alias{vrule}
\alias{vrule-package}
\title{Tools to validate data based on business validation rules}
\description{
Provides tools to validate data based on business validation rules.
}
