% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_abstract.R
\docType{class}
\name{vrule_abstract}
\alias{vrule_abstract}
\title{Abstract validation rule}
\description{
Abstract validation rule

Abstract validation rule
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_abstract-new}{\code{vrule_abstract$new()}}
\item \href{#method-vrule_abstract-validate}{\code{vrule_abstract$validate()}}
\item \href{#method-vrule_abstract-getType}{\code{vrule_abstract$getType()}}
\item \href{#method-vrule_abstract-getCategory}{\code{vrule_abstract$getCategory()}}
\item \href{#method-vrule_abstract-getName}{\code{vrule_abstract$getName()}}
\item \href{#method-vrule_abstract-clone}{\code{vrule_abstract$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_abstract-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_abstract-new}{}}}
\subsection{Method \code{new()}}{
Initializes an abstract validation rule
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_abstract$new(..., type = c("ERROR", "WARNING"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{args}

\item{\code{type}}{the type of rule either \code{ERROR} or \code{WARNING}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_abstract-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_abstract-validate}{}}}
\subsection{Method \code{validate()}}{
Abstract method to validate data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_abstract$validate(value, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{value}

\item{\code{...}}{any other args}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a validation report, object of class \link{vrule_report}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_abstract-getType"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_abstract-getType}{}}}
\subsection{Method \code{getType()}}{
Get validation type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_abstract$getType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{ERROR} or \code{WARNING}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_abstract-getCategory"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_abstract-getCategory}{}}}
\subsection{Method \code{getCategory()}}{
Get validation rule category
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_abstract$getCategory()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the validation rule category
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_abstract-getName"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_abstract-getName}{}}}
\subsection{Method \code{getName()}}{
Get validation rule name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_abstract$getName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the validation rule name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_abstract-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_abstract-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_abstract$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
