% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_abstract_complex.R
\docType{class}
\name{vrule_abstract_complex}
\alias{vrule_abstract_complex}
\title{vrule_abstract_complex}
\description{
vrule_abstract_complex

vrule_abstract_complex
}
\section{Super class}{
\code{\link[vrule:vrule_abstract]{vrule::vrule_abstract}} -> \code{vrule_abstract_complex}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_abstract_complex-new}{\code{vrule_abstract_complex$new()}}
\item \href{#method-vrule_abstract_complex-validate}{\code{vrule_abstract_complex$validate()}}
\item \href{#method-vrule_abstract_complex-clone}{\code{vrule_abstract_complex$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getType"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getType'><code>vrule::vrule_abstract$getType()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_abstract_complex-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_abstract_complex-new}{}}}
\subsection{Method \code{new()}}{
Initializes an abstract complex validation rule
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_abstract_complex$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{args}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_abstract_complex-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_abstract_complex-validate}{}}}
\subsection{Method \code{validate()}}{
Abstract method to validate data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_abstract_complex$validate(value, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{value}

\item{\code{...}}{any other args}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a validation report, object of class \link{vrule_report}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_abstract_complex-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_abstract_complex-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_abstract_complex$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
