% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_op.R
\docType{class}
\name{vrule_and}
\alias{vrule_and}
\title{vrule_and}
\description{
vrule_and

vrule_and
}
\section{Super classes}{
\code{vrule::vrule_abstract} -> \code{\link[vrule:vrule_op]{vrule::vrule_op}} -> \code{vrule_and}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_and-new}{\code{vrule_and$new()}}
\item \href{#method-vrule_and-validate}{\code{vrule_and$validate()}}
\item \href{#method-vrule_and-clone}{\code{vrule_and$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_and-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_and-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_and$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_and-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_and-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_and$validate(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_and-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_and-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_and$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
