% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_spatial.R
\docType{class}
\name{vrule_coordinate}
\alias{vrule_coordinate}
\alias{vrule_coordinates}
\title{vrule_coordinate}
\description{
vrule_coordinate

vrule_coordinate
}
\section{Super classes}{
\code{vrule::vrule_abstract} -> \code{vrule::vrule_abstract_simple} -> \code{\link[vrule:vrule_datatype]{vrule::vrule_datatype}} -> \code{\link[vrule:vrule_numeric]{vrule::vrule_numeric}} -> \code{vrule_coordinates}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_coordinates-new}{\code{vrule_coordinates$new()}}
\item \href{#method-vrule_coordinates-validate}{\code{vrule_coordinates$validate()}}
\item \href{#method-vrule_coordinates-clone}{\code{vrule_coordinates$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getType"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getType'><code>vrule::vrule_abstract$getType()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_coordinates-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_coordinates-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_coordinates$new(
  coordinate_type = c("latitude", "longitude"),
  na_allowed = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_coordinates-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_coordinates-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_coordinates$validate(value, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_coordinates-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_coordinates-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_coordinates$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
