% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_crossfield.R
\docType{class}
\name{vrule_cross_column}
\alias{vrule_cross_column}
\title{vrule_cross_column}
\description{
vrule_cross_column

vrule_cross_column
}
\section{Super classes}{
\code{vrule::vrule_abstract} -> \code{vrule::vrule_abstract_complex} -> \code{vrule_cross_column}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_cross_column-new}{\code{vrule_cross_column$new()}}
\item \href{#method-vrule_cross_column-validate}{\code{vrule_cross_column$validate()}}
\item \href{#method-vrule_cross_column-clone}{\code{vrule_cross_column$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_cross_column-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_cross_column-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_cross_column$new(operator, expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_cross_column-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_cross_column-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_cross_column$validate(value, row)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_cross_column-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_cross_column-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_cross_column$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
