% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_crossfield.R
\docType{class}
\name{vrule_crossfield_validator}
\alias{vrule_crossfield_validator}
\title{vrule_crossfield_validator}
\description{
vrule_crossfield_validator

vrule_crossfield_validator
}
\section{Super classes}{
\code{vrule::vrule_abstract} -> \code{vrule::vrule_abstract_complex} -> \code{vrule_crossfield_validator}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_crossfield_validator-new}{\code{vrule_crossfield_validator$new()}}
\item \href{#method-vrule_crossfield_validator-validate}{\code{vrule_crossfield_validator$validate()}}
\item \href{#method-vrule_crossfield_validator-clone}{\code{vrule_crossfield_validator$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_crossfield_validator-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_crossfield_validator-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_crossfield_validator$new(operator, expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_crossfield_validator-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_crossfield_validator-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_crossfield_validator$validate(value, row)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_crossfield_validator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_crossfield_validator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_crossfield_validator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
