% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_time.R
\docType{class}
\name{vrule_date}
\alias{vrule_date}
\title{vrule_date}
\description{
vrule_date

vrule_date
}
\section{Super classes}{
\code{vrule::vrule_abstract} -> \code{vrule::vrule_abstract_simple} -> \code{\link[vrule:vrule_datatype]{vrule::vrule_datatype}} -> \code{\link[vrule:vrule_character]{vrule::vrule_character}} -> \code{vrule_date}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_date-new}{\code{vrule_date$new()}}
\item \href{#method-vrule_date-validate}{\code{vrule_date$validate()}}
\item \href{#method-vrule_date-clone}{\code{vrule_date$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_date-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_date-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_date$new(na_allowed = FALSE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_date-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_date-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_date$validate(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_date-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_date-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_date$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
