% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_threshold.R
\docType{class}
\name{vrule_date_threshold}
\alias{vrule_date_threshold}
\title{vrule_date_threshold}
\description{
vrule_date_threshold

vrule_date_threshold
}
\section{Super classes}{
\code{vrule::vrule_abstract} -> \code{\link[vrule:vrule_operator_binary]{vrule::vrule_operator_binary}} -> \code{\link[vrule:vrule_operator_relational]{vrule::vrule_operator_relational}} -> \code{vrule_date_threshold}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_date_threshold-new}{\code{vrule_date_threshold$new()}}
\item \href{#method-vrule_date_threshold-validate}{\code{vrule_date_threshold$validate()}}
\item \href{#method-vrule_date_threshold-clone}{\code{vrule_date_threshold$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_date_threshold-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_date_threshold-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_date_threshold$new(operator, threshold)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_date_threshold-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_date_threshold-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_date_threshold$validate(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_date_threshold-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_date_threshold-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_date_threshold$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
