% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_datatype.R
\docType{class}
\name{vrule_double}
\alias{vrule_double}
\title{vrule_double}
\description{
vrule_double

vrule_double
}
\section{Super classes}{
\code{vrule::vrule_abstract} -> \code{vrule::vrule_abstract_simple} -> \code{\link[vrule:vrule_datatype]{vrule::vrule_datatype}} -> \code{vrule_double}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_double-new}{\code{vrule_double$new()}}
\item \href{#method-vrule_double-validate}{\code{vrule_double$validate()}}
\item \href{#method-vrule_double-clone}{\code{vrule_double$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_double-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_double-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_double$new(na_allowed = FALSE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_double-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_double-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_double$validate(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_double-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_double-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_double$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
