% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_if.R
\docType{class}
\name{vrule_if}
\alias{vrule_if}
\title{vrule_if}
\description{
vrule_if

vrule_if
}
\section{Super classes}{
\code{vrule::vrule_abstract} -> \code{vrule::vrule_abstract_complex} -> \code{vrule_if}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_if-new}{\code{vrule_if$new()}}
\item \href{#method-vrule_if-validate}{\code{vrule_if$validate()}}
\item \href{#method-vrule_if-clone}{\code{vrule_if$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_if-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_if-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_if$new(if_condition, then_apply = list(), else_apply = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_if-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_if-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_if$validate(value, row)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_if-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_if-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_if$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
