% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_datatype.R
\docType{class}
\name{vrule_logical_like}
\alias{vrule_logical_like}
\title{vrule_logical_like}
\description{
vrule_logical_like

vrule_logical_like
}
\section{Super classes}{
\code{vrule::vrule_abstract} -> \code{vrule::vrule_abstract_simple} -> \code{vrule_logical_like}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_logical_like-new}{\code{vrule_logical_like$new()}}
\item \href{#method-vrule_logical_like-validate}{\code{vrule_logical_like$validate()}}
\item \href{#method-vrule_logical_like-clone}{\code{vrule_logical_like$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getType"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getType'><code>vrule::vrule_abstract$getType()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_logical_like-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_logical_like-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_logical_like$new(na_allowed = FALSE, allowed_values = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_logical_like-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_logical_like-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_logical_like$validate(value, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_logical_like-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_logical_like-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_logical_like$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
