% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_spatial.R
\docType{class}
\name{vrule_lonlat_in_shape}
\alias{vrule_lonlat_in_shape}
\title{TRIAL vrule_lonlat_in_shape}
\description{
TRIAL vrule_lonlat_in_shape

TRIAL vrule_lonlat_in_shape
}
\section{Super classes}{
\code{vrule::vrule_abstract} -> \code{vrule::vrule_abstract_complex} -> \code{vrule_lonlat_in_shape}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_lonlat_in_shape-new}{\code{vrule_lonlat_in_shape$new()}}
\item \href{#method-vrule_lonlat_in_shape-validate}{\code{vrule_lonlat_in_shape$validate()}}
\item \href{#method-vrule_lonlat_in_shape-clone}{\code{vrule_lonlat_in_shape$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getType"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getType'><code>vrule::vrule_abstract$getType()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_lonlat_in_shape-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_lonlat_in_shape-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_lonlat_in_shape$new(
  shape,
  lon_field = "long",
  lat_field = "lat",
  na_allowed = FALSE,
  crs = 4326,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_lonlat_in_shape-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_lonlat_in_shape-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_lonlat_in_shape$validate(value, row, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_lonlat_in_shape-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_lonlat_in_shape-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_lonlat_in_shape$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
