% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_mapping.R
\docType{class}
\name{vrule_mapping}
\alias{vrule_mapping}
\title{vrule_mapping}
\description{
vrule_mapping

vrule_mapping
}
\section{Super classes}{
\code{vrule::vrule_abstract} -> \code{vrule::vrule_abstract_simple} -> \code{vrule_mapping}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_mapping-new}{\code{vrule_mapping$new()}}
\item \href{#method-vrule_mapping-validate}{\code{vrule_mapping$validate()}}
\item \href{#method-vrule_mapping-clone}{\code{vrule_mapping$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_mapping-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_mapping-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_mapping$new(
  ref_data_url = NULL,
  ref_source_term = NULL,
  ref_target_term = NULL,
  ref_meta_url = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_mapping-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_mapping-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_mapping$validate(value, target, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_mapping-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_mapping-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_mapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
