% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_time.R
\docType{class}
\name{vrule_month}
\alias{vrule_month}
\title{vrule_month (allows 1-12 or 01-12) could include a specifier to only allow one type of format}
\description{
vrule_month (allows 1-12 or 01-12) could include a specifier to only allow one type of format

vrule_month (allows 1-12 or 01-12) could include a specifier to only allow one type of format
}
\section{Super classes}{
\code{vrule::vrule_abstract} -> \code{vrule::vrule_abstract_simple} -> \code{\link[vrule:vrule_datatype]{vrule::vrule_datatype}} -> \code{\link[vrule:vrule_integer]{vrule::vrule_integer}} -> \code{vrule_month}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_month-new}{\code{vrule_month$new()}}
\item \href{#method-vrule_month-validate}{\code{vrule_month$validate()}}
\item \href{#method-vrule_month-clone}{\code{vrule_month$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getType"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getType'><code>vrule::vrule_abstract$getType()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_month-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_month-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_month$new(na_allowed = FALSE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_month-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_month-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_month$validate(value, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_month-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_month-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_month$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
