% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_datatype.R
\docType{class}
\name{vrule_numeric}
\alias{vrule_numeric}
\title{vrule_numeric}
\description{
vrule_numeric

vrule_numeric
}
\section{Super classes}{
\code{\link[vrule:vrule_abstract]{vrule::vrule_abstract}} -> \code{\link[vrule:vrule_abstract_simple]{vrule::vrule_abstract_simple}} -> \code{\link[vrule:vrule_datatype]{vrule::vrule_datatype}} -> \code{vrule_numeric}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_numeric-new}{\code{vrule_numeric$new()}}
\item \href{#method-vrule_numeric-validate}{\code{vrule_numeric$validate()}}
\item \href{#method-vrule_numeric-clone}{\code{vrule_numeric$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getType"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getType'><code>vrule::vrule_abstract$getType()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_numeric-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_numeric-new}{}}}
\subsection{Method \code{new()}}{
Initializes a numeric data validation rule
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_numeric$new(na_allowed = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{na_allowed}}{is NA allowed?}

\item{\code{...}}{any other arg}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_numeric-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_numeric-validate}{}}}
\subsection{Method \code{validate()}}{
Method to validate data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_numeric$validate(value, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{value}

\item{\code{...}}{any other args}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a validation report, object of class \link{vrule_report}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_numeric-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_numeric-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_numeric$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
