% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_op.R
\docType{class}
\name{vrule_op}
\alias{vrule_op}
\title{vrule_op}
\description{
vrule_op

vrule_op
}
\section{Super class}{
\code{vrule::vrule_abstract} -> \code{vrule_op}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_op-new}{\code{vrule_op$new()}}
\item \href{#method-vrule_op-validate}{\code{vrule_op$validate()}}
\item \href{#method-vrule_op-clone}{\code{vrule_op$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_op-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_op-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_op$new(operator, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_op-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_op-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_op$validate(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_op-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_op-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_op$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
