% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_operator_binary.R
\docType{class}
\name{vrule_operator_binary}
\alias{vrule_operator_binary}
\title{vrule_operator_binary}
\description{
vrule_operator_binary

vrule_operator_binary
}
\section{Super class}{
\code{vrule::vrule_abstract} -> \code{vrule_operator_binary}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_operator_binary-new}{\code{vrule_operator_binary$new()}}
\item \href{#method-vrule_operator_binary-validate}{\code{vrule_operator_binary$validate()}}
\item \href{#method-vrule_operator_binary-clone}{\code{vrule_operator_binary$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_operator_binary-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_operator_binary-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_operator_binary$new(operator, expr, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_operator_binary-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_operator_binary-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_operator_binary$validate(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_operator_binary-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_operator_binary-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_operator_binary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
