% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_operator_logical.R
\docType{class}
\name{vrule_operator_or}
\alias{vrule_operator_or}
\title{vrule_operator_or}
\description{
vrule_operator_or

vrule_operator_or
}
\section{Super classes}{
\code{vrule::vrule_abstract} -> \code{\link[vrule:vrule_operator_binary]{vrule::vrule_operator_binary}} -> \code{\link[vrule:vrule_operator_logical]{vrule::vrule_operator_logical}} -> \code{vrule_operator_or}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_operator_or-new}{\code{vrule_operator_or$new()}}
\item \href{#method-vrule_operator_or-validate}{\code{vrule_operator_or$validate()}}
\item \href{#method-vrule_operator_or-clone}{\code{vrule_operator_or$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getCategory"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getCategory'><code>vrule::vrule_abstract$getCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vrule" data-topic="vrule_abstract" data-id="getName"><a href='../../vrule/html/vrule_abstract.html#method-vrule_abstract-getName'><code>vrule::vrule_abstract$getName()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_operator_or-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_operator_or-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_operator_or$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_operator_or-validate"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_operator_or-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_operator_or$validate(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_operator_or-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_operator_or-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_operator_or$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
