% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrule_report.R
\docType{class}
\name{vrule_report}
\alias{vrule_report}
\title{vrule_report}
\description{
vrule_report

vrule_report
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vrule_report-new}{\code{vrule_report$new()}}
\item \href{#method-vrule_report-clone}{\code{vrule_report$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_report-new"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_report-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_report$new(valid = NULL, report = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vrule_report-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vrule_report-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vrule_report$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
