% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spec.R
\docType{class}
\name{format_spec}
\alias{format_spec}
\title{format_spec}
\description{
format_spec

format_spec
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-format_spec-new}{\code{format_spec$new()}}
\item \href{#method-format_spec-validateStructure}{\code{format_spec$validateStructure()}}
\item \href{#method-format_spec-validateContent}{\code{format_spec$validateContent()}}
\item \href{#method-format_spec-validate}{\code{format_spec$validate()}}
\item \href{#method-format_spec-validate_and_display_as_handsontable}{\code{format_spec$validate_and_display_as_handsontable()}}
\item \href{#method-format_spec-clone}{\code{format_spec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-new"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$new(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-validateStructure"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-validateStructure}{}}}
\subsection{Method \code{validateStructure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$validateStructure(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-validateContent"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-validateContent}{}}}
\subsection{Method \code{validateContent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$validateContent(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-validate"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$validate(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-validate_and_display_as_handsontable"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-validate_and_display_as_handsontable}{}}}
\subsection{Method \code{validate_and_display_as_handsontable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$validate_and_display_as_handsontable(data, use_css_classes = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
