% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spec.R
\docType{class}
\name{format_spec}
\alias{format_spec}
\title{format_spec}
\description{
format_spec

format_spec
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-format_spec-new}{\code{format_spec$new()}}
\item \href{#method-format_spec-setName}{\code{format_spec$setName()}}
\item \href{#method-format_spec-setURN}{\code{format_spec$setURN()}}
\item \href{#method-format_spec-setTitle}{\code{format_spec$setTitle()}}
\item \href{#method-format_spec-addColumnSpec}{\code{format_spec$addColumnSpec()}}
\item \href{#method-format_spec-getColumnSpecByName}{\code{format_spec$getColumnSpecByName()}}
\item \href{#method-format_spec-getColumnSpecByURN}{\code{format_spec$getColumnSpecByURN()}}
\item \href{#method-format_spec-getColumnSpecByAlias}{\code{format_spec$getColumnSpecByAlias()}}
\item \href{#method-format_spec-getColumnSpec}{\code{format_spec$getColumnSpec()}}
\item \href{#method-format_spec-validateStructure}{\code{format_spec$validateStructure()}}
\item \href{#method-format_spec-validateContent}{\code{format_spec$validateContent()}}
\item \href{#method-format_spec-validate}{\code{format_spec$validate()}}
\item \href{#method-format_spec-validate_and_display_as_handsontable}{\code{format_spec$validate_and_display_as_handsontable()}}
\item \href{#method-format_spec-clone}{\code{format_spec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-new"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$new(json = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-setName"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-setName}{}}}
\subsection{Method \code{setName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$setName(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-setURN"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-setURN}{}}}
\subsection{Method \code{setURN()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$setURN(urn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-setTitle"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-setTitle}{}}}
\subsection{Method \code{setTitle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$setTitle(title)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-addColumnSpec"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-addColumnSpec}{}}}
\subsection{Method \code{addColumnSpec()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$addColumnSpec(column_spec)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-getColumnSpecByName"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-getColumnSpecByName}{}}}
\subsection{Method \code{getColumnSpecByName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$getColumnSpecByName(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-getColumnSpecByURN"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-getColumnSpecByURN}{}}}
\subsection{Method \code{getColumnSpecByURN()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$getColumnSpecByURN(urn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-getColumnSpecByAlias"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-getColumnSpecByAlias}{}}}
\subsection{Method \code{getColumnSpecByAlias()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$getColumnSpecByAlias(alias)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-getColumnSpec"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-getColumnSpec}{}}}
\subsection{Method \code{getColumnSpec()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$getColumnSpec(column)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-validateStructure"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-validateStructure}{}}}
\subsection{Method \code{validateStructure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$validateStructure(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-validateContent"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-validateContent}{}}}
\subsection{Method \code{validateContent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$validateContent(
  data,
  method = "grid_mapply",
  parallel = FALSE,
  cl = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-validate"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$validate(
  data,
  method = "grid_mapply",
  parallel = FALSE,
  cl = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-validate_and_display_as_handsontable"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-validate_and_display_as_handsontable}{}}}
\subsection{Method \code{validate_and_display_as_handsontable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$validate_and_display_as_handsontable(
  data,
  method = "matrix",
  parallel = parallel,
  cl = cl,
  use_css_classes = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_spec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-format_spec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{format_spec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
